% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclate.R
\name{hclate}
\alias{hclate}
\title{hclate: Testing for Homogeneous Conditional Local Average Treatment Effetcs}
\usage{
hclate(out, delta, treat, inst, xvector, xpscore, b)
}
\arguments{
\item{out}{vector containing the outcome of interest}

\item{delta}{vector containing the censoring indicator (1 if observed, 0 if censored)}

\item{treat}{vector containing the treatment indicator (1 if treated, 0 if control)}

\item{inst}{vector containing the binary instrument}

\item{xvector}{matrix (or data frame) containing the conditioning covariates}

\item{xpscore}{matrix (or data frame) containing the covariates (and their
transformations) to be included in the propensity score estimation}

\item{b}{number of bootstrap draws}
}
\value{
a list containing the Kolmogorov-Smirnov test statistic (kstest),
       the Cramer-von Mises test statistic (cvmtest), and their associated
       bootstrapped p-values, pvks and pvcvm, respectively.
}
\description{
\emph{hclate} computes Kolmogorov-Smirnov and Cramer-von Mises type tests
for the null hypothesis of homogeneous conditional local average treatment effects.
The test is suitable for both censored and uncensored outcomes, and relies on
the availability of a binary instrumental variable that satisfies additional assumptions.
For details of the testing procedure, see Sant'Anna (2016b),'Nonparametric Tests for
Treatment Effect Heterogeneity with Censored data'.
}

