% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcate.R
\name{zcate}
\alias{zcate}
\title{zcate: Testing for Zero Conditional Average Treatment Effects}
\usage{
zcate(out, delta, treat, xvector, xpscore, b, cores = 1)
}
\arguments{
\item{out}{vector containing the outcome of interest}

\item{delta}{vector containing the censoring indicator (1 if observed, 0 if censored)}

\item{treat}{vector containing the treatment indicator (1 if treated, 0 if control)}

\item{xvector}{matrix (or data frame) containing the conditioning covariates}

\item{xpscore}{matrix (or data frame) containing the covariates (and their
transformations) to be included in the propensity score estimation}

\item{b}{number of bootstrap draws}

\item{cores}{number of cores to use during the bootstrap (default is 1).
If cores>1, the bootstrap is conducted using parLapply, instead
of lapply type call.}
}
\value{
a list containing the Kolmogorov-Smirnov test statistic (kstest),
       the Cramer-von Mises test statistic (cvmtest), and their associated
       bootstrapped p-values, pvks and pvcvm, respectively.
}
\description{
\emph{zcate} computes Kolmogorov-Smirnov and Cramer-von Mises type tests
for the null hypothesis of zero conditional average treatment effects.
The test is suitable for both censored and uncensored outcomes, and relies on
the unconfoundedness assumption. For details of the testing procedure, see
Sant'Anna (2016b),'Nonparametric Tests for Treatment Effect Heterogeneity with
Censored data'.
}

